session_supervisor: carve agent_config.rs (pure move)
